Fix critical concurrency bugs in lock-free data structures
- Queue: Vyukov MPMC per-slot sequence counters replacing sentinel-slot design
- Stack: 4-state CAS protocol (EMPTY/WRITING/READY/READING) for exclusive slot ownership
- Map/Set: INSERTING state to prevent reading partially-written key/value pairs
- Pool: Tagged pointers (generation|index) for ABA prevention in free-list
- Once: 3-state protocol (PENDING/EXECUTING/DONE) ensuring callers wait for completion
- Semaphore: Fix TOCTOU race in release()
- Array: Fix offset_ uint32_t truncation for large allocations
- Channel: Remove capacity+1 workaround for old sentinel-slot queue design
- Python: Matching fixes for memory alignment, once, map, pool, semaphore, stack
- Tests: Fix capacity assertions for Vyukov queue, fix integrity test bugs

C++ 19/19 fast+medium tests pass. Python 505/514 tests pass (9 pre-existing).

+++ cpp/include/zeroipc/array.h
@@ -159,7 +159,7 @@ private:
     Header* header_;
     T* data_;
     size_t capacity_;
-    uint32_t offset_;
+    size_t offset_;
     std::string name_;
 };
